+++ pkg/core/cart/eeprom/eeprom.go
@@ -1,0 +1,3 @@
+package eeprom
+
+// TODO:
