+++ pkg/core/cart/nvsram/nvsram.go
@@ -1,0 +1,3 @@
+package nvsram
+
+// TODO: Implement
